+++ samples/templates/default-azuredeploy.json
@@ -202,7 +202,7 @@
             "DicomFunctions__DurableTask__ConnectionName": "AzureWebJobsStorage",
             "DicomServer__Features__EnableOhifViewer": "[parameters('deployOhifViewer')]",
             "DicomServer__Features__EnableExtendedQueryTags": "true",
-            "DicomServer__Features__SupportNewBlobFormatForNewService": "true",
+            "DicomServer__Services__BlobMigration__FormatType": "2",
             "DicomServer__Security__Enabled": "[variables('securityAuthenticationEnabled')]",
             "DicomServer__Security__Authentication__Authority": "[parameters('securityAuthenticationAuthority')]",
             "DicomServer__Security__Authentication__Audience": "[parameters('securityAuthenticationAudience')]",
