+++ src/Microsoft.Health.Dicom.Blob.UnitTests/DicomFileNameWithUidTests.cs
@@ -3,12 +3,12 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using Microsoft.Health.Dicom.Core.Features.Common;
+using Microsoft.Health.Dicom.Blob.Features.Storage;
 using Microsoft.Health.Dicom.Core.Features.Model;
 using Microsoft.Health.Dicom.Tests.Common;
 using Xunit;
 
-namespace Microsoft.Health.Dicom.Core.UnitTests.Features.Common;
+namespace Microsoft.Health.Dicom.Blob.UnitTests.Features.Common;
 
 public class DicomFileNameWithUidTests
 {
