+++ src/Microsoft.Health.Dicom.Core/Microsoft.Health.Dicom.Core.csproj
@@ -9,7 +9,6 @@
     <PackageReference Include="Ensure.That" Version="10.1.0" />
     <PackageReference Include="fo-dicom" Version="$(FoDicomVersion)" />
     <PackageReference Include="fo-dicom.Codecs" Version="5.1.0" />
-    <PackageReference Include="HashDepot" Version="2.0.3" />
     <PackageReference Include="MediatR" Version="9.0.0" />
     <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="9.0.0" />
     <PackageReference Include="Microsoft.ApplicationInsights" Version="2.20.0" />
