+++ src/Microsoft.Health.Dicom.Web/appsettings.Development.json
@@ -15,9 +15,12 @@
   },
   "DicomServer": {
     "Features": {
-      "EnableUpsRs": true,
-      "EnableDualWrite": false,
-      "SupportNewBlobFormatForNewService": true
+      "EnableUpsRs": true
+    },
+    "Services": {
+      "BlobMigration": {
+        "FormatType": 0
+      }
     }
   },
   "Logging": {
