+++ src/Microsoft.Health.Dicom.Web/appsettings.json
@@ -77,6 +77,9 @@
       },
       "ExtendedQueryTag": {
         "MaxAllowedCount": 128
+      },
+      "BlobMigration": {
+        "FormatType": 0
       }
     },
     "Audit": {
